erpt1610: use bootstrap-salt.sh script provided by salt-cloud
change it to use our archive && version.
Existing script cannot be used anymore in 2014.1.10 because salt-cloud
use new temp_dir, and it do not pass that var to any script than one
with whose name is bootstrap-salt.
Maintaining a separate script for bootstrap is just add more work,
since upstream often get changes. So use upstream one and patch it to
our way.