erp2960: reorganize top file
always use top file in base, it will authoritative in all env,
fix {{ base }} -> base, it is not a jinja2 var.

Top file need to be moved to other path because if it is at the high
of the dir hierachy, user can access states in subdir such as
states.repo1.apt (relative path to top /srv/salt). Which might not
cause harm but may confuse user