erp3856: allow appending last resorts when resolving DNS
this is useful when using an internal DNS server, which act only
as authoritative server, when all prior server failed to resolve
DNS, it will try that server. Because as it acts as authoritative
only, it will log many error when used to resolve DNS, let other
prior name servers do that.